refactor(agents): own the agent-context record and its merge rules
Keep the persisted agent model and its pure merge policy with the agent
context domain while the store retains only sidecar I/O choreography.
