+++ crates/rimz/src/agents/context.rs
@@ -23,6 +23,10 @@ use crate::ids::{AgentSessionId, MessageId};
 use super::LocalContextRefresh;
 use super::definition::AgentSpec;
 
+mod record;
+
+pub use record::AgentContextRecord;
+
 /// One rich-context reading paired with the provider session that owns it.
 #[derive(Clone, Debug, PartialEq)]
 pub struct ContextObservation {
@@ -242,7 +246,7 @@ pub struct SessionContextInput<'a> {
     pub server_url: Option<&'a str>,
     /// The sidecar as last written, so an adapter can throttle its expensive
     /// read and diff against what is already stored.
-    pub prior: Option<&'a crate::store::agent_context::AgentContextRecord>,
+    pub prior: Option<&'a AgentContextRecord>,
     /// The shared price book, for adapters that fold cost while refreshing.
     pub pricing_cache_path: &'a Path,
     /// This session's warm broker socket, for adapters that host one.
