+++ crates/rimz/src/store/agent_context.rs
@@ -21,72 +21,15 @@
 use std::path::PathBuf;
 
 use jiff::Timestamp;
-use serde::{Deserialize, Serialize};
 
-use crate::agents::context::{AgentContext, AgentTurnError, merge_session_usage};
+use crate::agents::context::{AgentContext, AgentContextRecord, AgentTurnError};
 use crate::agents::{
-    AgentCost, AgentSpec, AgentState, AgentStatus, AgentTokenUsage, LocalContextRefresh,
-    LocalSpendFold, TranscriptStat,
+    AgentSpec, AgentState, AgentStatus, LocalContextRefresh, LocallyPricedTurnCost,
 };
 use crate::disk::atomic;
 use crate::disk::paths::RuntimePaths;
-use crate::ids::{AgentKind, AgentSessionId, MessageId};
+use crate::ids::MessageId;
 use crate::store::sidecar;
-
-/// A session's context sidecar: the normalized record plus the
-/// `(kind, agent_id)` it is filed under, so a read can confirm the key — and
-/// shrug off a digest collision — instead of trusting the filename.
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub struct AgentContextRecord {
-    pub kind: AgentKind,
-    pub agent_id: AgentSessionId,
-    pub context: AgentContext,
-    /// When app-server/account-scoped context was last observed. Local transcript
-    /// pushes bump `context.observed_at`, so app-server throttles use this stamp
-    /// instead of the whole-record freshness.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub rate_limits_observed_at: Option<Timestamp>,
-    /// When a rich-context transport last wrote display-only metadata that is
-    /// not rate-limit/account data. Local token/cost pushes bump
-    /// `context.observed_at`, so rich-context throttles use this stamp instead
-    /// of whole-record freshness.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub rich_observed_at: Option<Timestamp>,
-    /// Transcript, rollout, or telemetry file used for the latest local context refresh.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub transcript_path: Option<String>,
-    /// Stat gate for [`Self::transcript_path`], letting high-frequency hooks skip
-    /// an unchanged tail without parsing it.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub transcript_stat: Option<TranscriptStat>,
-    /// Resumable per-request pricing state for the local transcript.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub spend_fold: Option<LocalSpendFold>,
-    /// Hook-priced live-session state. Private so only the idempotent merge can
-    /// advance the accumulator.
-    #[serde(default, skip_serializing_if = "LocallyPricedCostState::is_empty")]
-    locally_priced_cost: LocallyPricedCostState,
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-struct LocallyPricedCostState {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    last_turn_id: Option<String>,
-    #[serde(default, skip_serializing_if = "is_zero_f64")]
-    cumulative_usd: f64,
-    #[serde(default, skip_serializing_if = "std::ops::Not::not")]
-    owns_context_cost: bool,
-}
-
-impl LocallyPricedCostState {
-    fn is_empty(&self) -> bool {
-        self.last_turn_id.is_none() && self.cumulative_usd == 0.0 && !self.owns_context_cost
-    }
-}
-
-fn is_zero_f64(value: &f64) -> bool {
-    *value == 0.0
-}
 
 impl sidecar::SidecarRecord for AgentContextRecord {
     const FILE_PREFIX: &'static str = "ctx";
@@ -122,41 +65,7 @@ pub fn write(
 ) -> Result<(), atomic::AtomicErr> {
     let observed_at = context.observed_at;
     update_record(runtime, kind, agent_id, observed_at, |record, _| {
-        let mut context = context.clone();
-        let observed_cost = context.cost.is_some();
-        // Statusline producers own their observed fields, while lifecycle
-        // confirmation and local transcript folds own turn causality and the
-        // resolved transcript path. Preserve those independently-written fields
-        // across whole-context statusline refreshes.
-        context.turn_opened_by = record.context.turn_opened_by.clone();
-        if context.cost.is_none() {
-            context.cost.clone_from(&record.context.cost);
-        }
-        let prior_session_usage = record
-            .context
-            .tokens
-            .as_ref()
-            .and_then(|tokens| tokens.session_usage.clone());
-        if let Some(prior_session_usage) = prior_session_usage {
-            let tokens = context.tokens.get_or_insert_with(AgentTokenUsage::default);
-            merge_session_usage(&mut tokens.session_usage, Some(prior_session_usage));
-        }
-        let mut locally_priced_cost = record.locally_priced_cost.clone();
-        if observed_cost {
-            locally_priced_cost.owns_context_cost = false;
-        }
-        *record = AgentContextRecord {
-            kind: AgentKind::new_unchecked(kind),
-            agent_id: agent_id.into(),
-            context,
-            rate_limits_observed_at: None,
-            rich_observed_at: None,
-            transcript_path: record.transcript_path.clone(),
-            transcript_stat: None,
-            spend_fold: record.spend_fold.clone(),
-            locally_priced_cost,
-        };
-        true
+        record.apply_context_refresh(kind, agent_id, context.clone())
     })
     .map(|_| ())
 }
@@ -194,27 +103,7 @@ pub fn write_record(
         record.kind.as_str(),
         record.agent_id.as_str(),
         observed_at,
-        |current, existed| {
-            let mut next = record.clone();
-            let observed_cost =
-                next.context.cost.is_some() && !next.locally_priced_cost.owns_context_cost;
-            if next.context.cost.is_none() {
-                next.context.cost.clone_from(&current.context.cost);
-            }
-            if next.locally_priced_cost.is_empty() {
-                next.locally_priced_cost = current.locally_priced_cost.clone();
-            }
-            // Local folds advance only through merge_local_context under this lock;
-            // whole-record writers preserve the latest fold if their read raced it.
-            if existed {
-                next.spend_fold.clone_from(&current.spend_fold);
-            }
-            if observed_cost {
-                next.locally_priced_cost.owns_context_cost = false;
-            }
-            *current = next;
-            true
-        },
+        |current, existed| current.apply_fixture(record.clone(), existed),
     )
     .map(|_| ())
 }
@@ -233,7 +122,7 @@ pub fn update_record(
         &runtime.agent_context_dir,
         kind,
         agent_id,
-        || new_record(kind, agent_id, AgentContext::new(kind, observed_at)),
+        || AgentContextRecord::new(kind, agent_id, AgentContext::new(kind, observed_at)),
         update,
     )
 }
@@ -243,20 +132,6 @@ pub fn update_record(
 /// published bytes, not the last value a sidebar consumer happened to parse.
 pub fn read_one(runtime: &RuntimePaths, kind: &str, agent_id: &str) -> Option<AgentContextRecord> {
     sidecar::read_one(&runtime.agent_context_dir, kind, agent_id)
-}
-
-pub fn new_record(kind: &str, agent_id: &str, context: AgentContext) -> AgentContextRecord {
-    AgentContextRecord {
-        kind: AgentKind::new_unchecked(kind),
-        agent_id: agent_id.into(),
-        context,
-        rate_limits_observed_at: None,
-        rich_observed_at: None,
-        transcript_path: None,
-        transcript_stat: None,
-        spend_fold: None,
-        locally_priced_cost: LocallyPricedCostState::default(),
-    }
 }
 
 /// Merge transcript-derived local context into a sidecar record. Local refresh
@@ -269,20 +144,13 @@ pub fn merge_local_context(
     refresh: LocalContextRefresh,
     observed_at: Timestamp,
 ) -> Result<(), atomic::AtomicErr> {
-    let kind = definition.kind;
-    update_record(runtime, kind, agent_id, observed_at, |record, _| {
-        record.context.source = kind.to_owned();
-        let cost_replaced = !refresh.context.cost.is_keep();
-        refresh.context.apply(&mut record.context, definition);
-        if cost_replaced {
-            record.locally_priced_cost.owns_context_cost = false;
-        }
-        record.context.observed_at = observed_at;
-        refresh.spend_fold.apply(&mut record.spend_fold);
-        record.transcript_path = refresh.transcript_path;
-        record.transcript_stat = refresh.transcript_stat;
-        true
-    })
+    update_record(
+        runtime,
+        definition.kind,
+        agent_id,
+        observed_at,
+        |record, _| record.apply_local_refresh(definition, refresh, observed_at),
+    )
     .map(|_| ())
 }
 
@@ -298,59 +166,7 @@ pub fn merge_observed(
 ) -> Result<bool, atomic::AtomicErr> {
     let observed_at = context.observed_at;
     update_record(runtime, kind, agent_id, observed_at, |record, _| {
-        let mut changed = false;
-        macro_rules! merge_optional {
-            ($field:ident) => {
-                if let Some(value) = context.$field
-                    && record.context.$field.as_ref() != Some(&value)
-                {
-                    record.context.$field = Some(value);
-                    changed = true;
-                }
-            };
-        }
-        merge_optional!(session_name);
-        merge_optional!(session_preview);
-        merge_optional!(model_display_name);
-        merge_optional!(thinking_enabled);
-        merge_optional!(output_style);
-        merge_optional!(vim_mode);
-        merge_optional!(agent_version);
-        merge_optional!(exceeds_200k_tokens);
-        merge_optional!(pr);
-        merge_optional!(account);
-        merge_optional!(turn_error);
-        merge_optional!(settle);
-        merge_optional!(model_id);
-        merge_optional!(effort);
-        if let Some(rate_limits) = context.rate_limits
-            && record.context.rate_limits.as_ref() != Some(&rate_limits)
-        {
-            record.context.rate_limits = Some(rate_limits);
-            record.rate_limits_observed_at = Some(observed_at);
-            changed = true;
-        }
-        if let Some(tokens) = context.tokens {
-            changed |= merge_observed_tokens(&mut record.context.tokens, tokens);
-        }
-        if let Some(cost) = context.cost
-            && let Some(total_cost_usd) = cost.total_cost_usd
-        {
-            let prior_total_cost = record
-                .context
-                .cost
-                .as_ref()
-                .and_then(|cost| cost.total_cost_usd);
-            if prior_total_cost.is_none_or(|prior| total_cost_usd >= prior) {
-                changed |= merge_observed_cost(&mut record.context.cost, cost, total_cost_usd);
-                record.locally_priced_cost.owns_context_cost = false;
-            }
-        }
-        if changed {
-            record.context.source = kind.to_owned();
-            record.context.observed_at = observed_at;
-        }
-        changed
+        record.merge_observed(kind, context, observed_at)
     })
 }
 
@@ -402,77 +218,15 @@ pub fn merge_locally_priced_cost(
     runtime: &RuntimePaths,
     kind: &str,
     agent_id: &str,
-    priced: &crate::agents::LocallyPricedTurnCost,
+    priced: &LocallyPricedTurnCost,
 ) -> Result<bool, atomic::AtomicErr> {
     if priced.turn_id.trim().is_empty() || !priced.cost_usd.is_finite() || priced.cost_usd <= 0.0 {
         return Ok(false);
     }
     let observed_at = Timestamp::now();
     update_record(runtime, kind, agent_id, observed_at, |record, _| {
-        if record.locally_priced_cost.last_turn_id.as_deref() == Some(priced.turn_id.as_str()) {
-            return false;
-        }
-        let cumulative = record.locally_priced_cost.cumulative_usd + priced.cost_usd;
-        if !cumulative.is_finite() || cumulative < 0.0 {
-            return false;
-        }
-        record.locally_priced_cost.last_turn_id = Some(priced.turn_id.clone());
-        record.locally_priced_cost.cumulative_usd = cumulative;
-        record.context.source = kind.to_owned();
-        record.context.observed_at = observed_at;
-        if record.locally_priced_cost.owns_context_cost || record.context.cost.is_none() {
-            record.locally_priced_cost.owns_context_cost = true;
-            let cost = record.context.cost.get_or_insert_with(AgentCost::default);
-            cost.total_cost_usd = Some(cumulative);
-        }
-        true
+        record.apply_locally_priced_turn(kind, priced, observed_at)
     })
-}
-
-fn merge_observed_tokens(prior: &mut Option<AgentTokenUsage>, incoming: AgentTokenUsage) -> bool {
-    let target = prior.get_or_insert_with(AgentTokenUsage::default);
-    let before = target.clone();
-    if incoming.context_window_size.is_some() {
-        target.context_window_size = incoming.context_window_size;
-    }
-    if incoming.used_percentage.is_some() {
-        target.used_percentage = incoming.used_percentage;
-    }
-    if incoming.remaining_percentage.is_some() {
-        target.remaining_percentage = incoming.remaining_percentage;
-    }
-    if incoming.current_context_tokens.is_some() {
-        target.current_context_tokens = incoming.current_context_tokens;
-    }
-    if let Some(current_usage) = incoming.current_usage {
-        target.current_usage = Some(current_usage);
-    }
-    crate::agents::context::merge_session_usage(&mut target.session_usage, incoming.session_usage);
-    *target != before
-}
-
-fn merge_observed_cost(
-    prior: &mut Option<AgentCost>,
-    incoming: AgentCost,
-    total_cost_usd: f64,
-) -> bool {
-    let target = prior.get_or_insert_with(AgentCost::default);
-    let before = target.clone();
-    target.total_cost_usd = Some(total_cost_usd);
-    target.coverage = incoming.coverage;
-    if incoming.total_duration_ms.is_some() {
-        target.total_duration_ms = incoming.total_duration_ms;
-    }
-    if incoming.total_api_duration_ms.is_some() {
-        target.total_api_duration_ms = incoming.total_api_duration_ms;
-    }
-    if incoming.total_lines_added.is_some() {
-        target.total_lines_added = incoming.total_lines_added;
-    }
-    if incoming.total_lines_removed.is_some() {
-        target.total_lines_removed = incoming.total_lines_removed;
-    }
-    *target != before
 }
 
 thread_local! {
