+++ crates/rimz/tests/integration/hooks.rs
@@ -1263,7 +1263,8 @@ fn cursor_concurrent_subagents_fold_independently_without_context_sidecars() {
     }));
     let mut context = rimz::agents::AgentContext::new("cursor", Timestamp::now());
     context.model_id = Some("parent-sidecar-model".to_owned());
-    let parent_context = rimz::store::agent_context::new_record("cursor", "cursor-parent", context);
+    let parent_context =
+        rimz::agents::context::AgentContextRecord::new("cursor", "cursor-parent", context);
     rimz::store::agent_context::write_record(&env.runtime_paths(), &parent_context)
         .expect("seed parent context");
     let parent_context_path =
@@ -1874,7 +1875,7 @@ fn codex_subagent_lifecycle_uses_child_agent_identity() {
     let mut context = rimz::agents::AgentContext::new("codex", Timestamp::now());
     context.model_id = Some("parent-model".to_owned());
     let parent_context =
-        rimz::store::agent_context::new_record("codex", "sess-codex-parent", context);
+        rimz::agents::context::AgentContextRecord::new("codex", "sess-codex-parent", context);
     rimz::store::agent_context::write_record(&env.runtime_paths(), &parent_context)
         .expect("seed parent context");
     let parent_context_path =
